drop fork-bundleAsync
