+++ packages/@expo/cli/src/start/server/middleware/resolveAssets.ts
@@ -3,7 +3,7 @@ import fs from 'fs/promises';
 import path from 'path';
 
 import { getAssetSchemasAsync } from '../../../api/getExpoSchema';
-import { BundleAssetWithFileHashes } from '../../../export/fork-bundleAsync';
+import { BundleAssetWithFileHashes } from '../../../export/saveAssets';
 import * as Log from '../../../log';
 import { fileExistsAsync } from '../../../utils/dir';
 import { CommandError } from '../../../utils/errors';
